requirements.txt
easyocr==1.7.1
filelock==3.15.1
fsspec==2024.6.0
imageio==2.34.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
mysql-connector-python==8.4.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.3.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
python-dotenv==1.0.1
PyYAML==6.0.1
scikit-image==0.24.0
scipy==1.13.1
setuptools==70.0.0
shapely==2.0.4
six==1.16.0
sympy==1.12.1
tifffile==2024.6.18
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
wheel==0.43.0
